修正因为添加-Werror编译选项暴露的代码问题

+++ ps_protocol/ps_protocol_as3.cpp
@@ -9,7 +9,7 @@ int protocol_as3_recv(int socketfd,char* buf)
 	while(true)
 	{
 		received = 0;
-		received = recv(socketfd,buf + offset,1,0);
+		received = (int)recv(socketfd,buf + offset,1,0);
 		if(received <= 0)
 		{
 			return -1;
