+++ ps_socket/ps_socket_transfer.cpp
@@ -4,13 +4,13 @@
 int socket_recv_by_size(int socketfd,char* buf,size_t nbytes)
 {
 	int received = 0;
-	int receiving = nbytes;
+	int receiving = (int)nbytes;
 	int offset = 0;
 
 	while(receiving > 0)
 	{
 		received = 0;
-		received = recv(socketfd,buf + offset,receiving,0);
+		received = (int)recv(socketfd,buf + offset,receiving,0);
 		if(received <= 0)
 		{
 			return -1;
@@ -19,19 +19,19 @@ int socket_recv_by_size(int socketfd,char* buf,size_t nbytes)
 		offset = offset + received;
 	}
 
-	return nbytes;
+	return (int)nbytes;
 }
 
 int socket_send_by_size(int socketfd,const char* buf,size_t nbytes)
 {
 	int sent = 0;
-	int sending = nbytes;
+	int sending = (int)nbytes;
 	int offset = 0;
 
 	while(sending > 0)
 	{
 		sent = 0;
-		sent = send(socketfd,buf + offset,sending,0);
+		sent = (int)send(socketfd,buf + offset,sending,0);
 		if(sent <= 0)
 		{
 			return -1;
@@ -40,7 +40,7 @@ int socket_send_by_size(int socketfd,const char* buf,size_t nbytes)
 		offset = offset + sent;
 	}
 
-	return nbytes;
+	return (int)nbytes;
 }
